+++ fastopenapi/routers/__init__.py
@@ -4,6 +4,11 @@ class MissingRouter:
 
 
 try:
+    from fastopenapi.routers.aiohttp import AioHttpRouter
+except ModuleNotFoundError:
+    AioHttpRouter = MissingRouter
+
+try:
     from fastopenapi.routers.falcon import FalconRouter
 except ModuleNotFoundError:
     FalconRouter = MissingRouter
@@ -34,6 +39,7 @@ except ModuleNotFoundError:
     TornadoRouter = MissingRouter
 
 __all__ = [
+    "AioHttpRouter",
     "FalconRouter",
     "FlaskRouter",
     "QuartRouter",