+++ tests/flask/test_flask_integration.py
@@ -19,7 +19,7 @@ class TestFlaskIntegration:
 
         assert response.status_code == 500
         result = json.loads(response.text)
-        assert result["detail"] == "TEST ERROR"
+        assert result["error"]["message"] == "TEST ERROR"
 
     def test_get_item(self, client):
         """Test fetching an item by ID"""
@@ -31,15 +31,14 @@ class TestFlaskIntegration:
         assert result["name"] == "Item 1"
         assert result["description"] == "Description 1"
 
-    def test_get_item_unprocessable(self, client):
+    def test_get_item_bad_request(self, client):
         """Test fetching an item with an incorrect ID type"""
         response = client.get("/items/abc")
 
-        assert response.status_code == 422
+        assert response.status_code == 400
         result = json.loads(response.text)
-        assert result["detail"] == (
-            "Error casting parameter 'item_id' to <class 'int'>: "
-            "invalid literal for int() with base 10: 'abc'"
+        assert result["error"]["message"] == (
+            "Error parsing parameter 'item_id'. Must be a valid int"
         )
 
     def test_get_nonexistent_item(self, client):
@@ -74,7 +73,7 @@ class TestFlaskIntegration:
 
         assert response.status_code == 422
         result = json.loads(response.text)
-        assert "Validation error for parameter" in result["detail"]
+        assert "Validation error for parameter" in result["error"]["message"]
 
     def test_create_item_invalid_json(self, client):
         """Test creating an item with invalid JSON"""
@@ -86,7 +85,7 @@ class TestFlaskIntegration:
 
         assert response.status_code == 422
         result = json.loads(response.text)
-        assert "Validation error for parameter" in result["detail"]
+        assert "Validation error for parameter" in result["error"]["message"]
 
     def test_update_item(self, client):
         """Test updating an item"""