+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to FastOpenAPI are documented in this file.
 
 FastOpenAPI follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [0.5.0] - Unreleased
+
+### Added
+- `AioHttpRouter` for integration with the `AioHttp` framework
+- Class-level cache for model schemas
+- `response_errors` for routers
+- `error_handler` for standard error responses
+
 ## [0.4.0] - 2025-03-20
 
 ### Added